Update cipher.js
**CSS:**
Added mobile styling: output box leaves, the two remaining boxes stack vertically, wheel floats left of the settings buttons.
Removed unused colour variables.
Fixed the height and width so that the app fits onto the screen.
Re-aligned the header.
Put the buttons below the first box into a flex container and hid the copy button for use only on mobile.
Removed the rotate CSS as I altered the image to be the correct angle instead.

**JS:**
On the new mobile layout, I wanted to use just one box for input and output. To do so, I had to create a proxy input (like a memory) and change the cipher from automatically working on INPUT_TEXT after every settings change/encrypt click to working on the input proxy, which is only updated when the textarea receives input. This allowed me to take the input proxy value, run it through cipher, and then paste it into the input box directly.
I changed most of the functions to use ES6 arrow syntax. With most functions being a ternary or imperative command, this was light work but the result is clean, tidy code.
I added a couple of const functions to check whether the device was in mobile- or desktop-view and to set the output path accordingly.